[dad] -Fix: spin: convert fabs() argument to double (false alarm warning on openbsd)
git-svn-id: svn://repo.hu/pcb-rnd/trunk@24835 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/hid_dad_spin.c
@@ -212,12 +212,12 @@ static double get_step(pcb_hid_dad_spin_t *spin, pcb_hid_attribute_t *end, pcb_h
 
 	switch(spin->type) {
 		case PCB_DAD_SPIN_INT:
-			step = pow(10, floor(log10(fabs(end->default_val.int_value)) - 1.0));
+			step = pow(10, floor(log10(fabs((double)end->default_val.int_value)) - 1.0));
 			if (step < 1)
 				step = 1;
 			break;
 		case PCB_DAD_SPIN_DOUBLE:
-			step = pow(10, floor(log10(fabs(end->default_val.real_value)) - 1.0));
+			step = pow(10, floor(log10(fabs((double)end->default_val.real_value)) - 1.0));
 			break;
 		case PCB_DAD_SPIN_COORD:
 			if (spin->unit == NULL) {
